Removed the explicit GC call when Realm changes
Fix #1925
GC should not be forced to run

If references of RealmResults from async queries are kept, the async
query will be re-run when the Realm changes to keep it updated. But if
the all of the references to the RealmResults are freed, Realm has no
way to know until it gets GCed. In that case, the async query will still
be re-run which is wasting. We are checking if the RealmResults is no
longer used by keeping a WeakReference to it and check when Realm updated.

Calling GC before is to prevent those wasted re-run async queries. But
it makes user loose control of the timing of UI thread ---- GC could
block UI thread and user cannot do anything with it.